+++ encodings/fastlanes/src/bitpacking/compute/take.rs
@@ -17,7 +17,7 @@ use vortex_dtype::{
 use vortex_error::{VortexExpect as _, VortexResult};
 
 use super::chunked_indices;
-use crate::{BitPackedArray, BitPackedVTable, bitpack_compress};
+use crate::{BitPackedArray, BitPackedVTable, bitpack_decompress};
 
 /// assuming the buffer is already allocated (which will happen at most once) then unpacking
 /// all 1024 elements takes ~8.8x as long as unpacking a single element on an M2 Macbook Air.
@@ -108,7 +108,7 @@ fn take_primitive<T: NativePType + BitPacking, I: IntegerPType>(
                 // so we need to unpack each one individually
                 for &index in offset_chunk_iter.remainder() {
                     output.push(unsafe {
-                        bitpack_compress::unpack_single_primitive::<T>(packed, bit_width, index)
+                        bitpack_decompress::unpack_single_primitive::<T>(packed, bit_width, index)
                     });
                 }
             }
